Move isOnAzure into own package
This is so the function can be re-used by the integrationtests Jenkins
job, decoupled from the microservice deployment logic.
